refactor(types): toolbar
