+++ src/lib/builders/toolbar/types.ts
@@ -1,0 +1,25 @@
+import type { createToolbar } from './create';
+
+export type CreateToolbarArgs = {
+	loop?: boolean;
+	orientation?: 'horizontal' | 'vertical';
+};
+
+type SingleToolbarGroupRootArgs = {
+	type?: 'single';
+	value?: string | null;
+};
+
+type MultipleToolbarGroupRootProps = {
+	type: 'multiple';
+	value?: string[];
+};
+
+export type CreateToolbarGroupArgs = (
+	| SingleToolbarGroupRootArgs
+	| MultipleToolbarGroupRootProps
+) & {
+	disabled?: boolean;
+};
+
+export type CreateToolbarReturn = ReturnType<typeof createToolbar>;
